# api/openapi.yaml
openapi: 3.0.3
info:
  title: TimeBeacon Intelligent Time Tracking API
  description: |
    Production-ready API for intelligent time tracking with LLM-powered duration estimation,
    project matching, and automated data import from email, calendar, and document sources.
  version: 1.0.0
  contact:
    name: TimeBeacon API Support
    email: api@timebeacon.io
  license:
    name: Proprietary
servers:
  - url: https://api.timebeacon.io/v1
    description: Production server
  - url: http://localhost:3001/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /auth/google/callback:
    get:
      tags: [Authentication]
      summary: Handle Google OAuth callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Authentication successful
          content:
            text/html:
              schema:
                type: string

  # Time Entries
...
  /import/gmail:
    post:
      tags: [Data Import]
      summary: Import and process Gmail emails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GmailImportRequest'
      responses:
        '200':
          description: Gmail import completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResponse'
  /import/calendar:
    post:
      tags: [Data Import]
      summary: Import and process calendar events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarImportRequest'
      responses:
        '200':
          description: Calendar import completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResponse'
  /import/drive:
    post:
      tags: [Data Import]
      summary: Import and process Google Drive activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriveImportRequest'
      responses:
        '200':
          description: Drive import completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResponse'

  # Projects & Clients
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GmailImportRequest:
      type: object
      required: [start_date]
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        max_emails:
          type: integer
          default: 100
        query:
          type: string
          description: Gmail search query
    CalendarImportRequest:
      type: object
      required: [start_date]
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        calendar_id:
          type: string
          default: primary
    ImportResponse:
      type: object
      properties:
        imported_count:
          type: integer
        processed_count:
          type: integer
        failed_count:
          type: integer
        time_entries_created:
          type: array
          items:
            type: string
            format: uuid
        processing_time_ms:
          type: integer
        errors:
          type: array
          items:
            type: string

    # Standard Responses
...